Updated async paths to know which function to post-process
As requested, setup for any particular async call in fs-ext.cc will record
which operation is being done, so that the async post-processing in After()
can detect how many and which values to return as arguments to a callback.

Changes to flock() function test to check only one argument is being returned.